CS-229 Moved permission seeding from test data seeder to seeder.
