+++ csatar-plugins/csatar/controllers/seederdata/seeder.htm
@@ -33,6 +33,18 @@
                             echo $contact_value . '; ';
                         }
                     }
+                } elseif ($key == 'permissions') {
+                    echo '<ul>';
+                    foreach ($value as $key => $permissions) {
+                        if (!is_array($permissions)) {
+                            echo '<li>' . $key . ': ' . e(trans('csatar.csatar::lang.plugin.admin.permissions.' . $permissions)) . '</li>';
+                        } else {
+                            foreach($permissions as $permission) {
+                                echo '<li>' . $key . ': ' . e(trans('csatar.csatar::lang.plugin.admin.permissions.' . $permission)) . '</li>';
+                            }
+                        }
+                    }
+                    echo '</ul>';
                 }
                 echo '</li>';
             }
