+++ csatar-plugins/csatar/controllers/seederdata/test.htm
@@ -6,19 +6,6 @@
             foreach ($testData as $key => $value) {
                 if (array_key_exists(0, $value) && !is_array($value[0])) {
                     echo '<li>' . e(trans('csatar.csatar::lang.plugin.admin.' . $key . '.' . $key)) . ': ' . implode(', ', $value) . '</li>';
-                } elseif ($key == 'permissions') {
-                    echo '<li>' . e(trans('csatar.csatar::lang.plugin.admin.' . $key . '.' . $key)) . ': </li>';
-                    echo '<ul>';
-                    foreach ($value as $key => $permissions) {
-                        if (!is_array($permissions)) {
-                            echo '<li>' . $key . ': ' . e(trans('csatar.csatar::lang.plugin.admin.permissions.' . $permissions)) . '</li>';
-                        } else {
-                            foreach($permissions as $permission) {
-                                echo '<li>' . $key . ': ' . e(trans('csatar.csatar::lang.plugin.admin.permissions.' . $permission)) . '</li>';
-                            }
-                        }
-                    }
-                    echo '</ul>';
                 }
             }
         ?>
